requirements.txt
Jinja2==2.10
PyYAML==3.12